refactor(analysis): split content tag helpers
